pyproject.toml
[project]
name = "apel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "pydantic>=2.11.7",
    "python-telegram-bot[job-queue]>=22.3",
    "requests>=2.32.5",
    "ruff>=0.12.10",
]
